feature(backend): Add `jira_issue` Backend
Adding a new backend `jira_issue` which allows verifying
issues based on a status of a Jira issue.

Refactor `JiraConnection` class to be a singelton that may
be reused in backends that require access to the Jira instance.

Update documentation to include `jira_backend`.

Fix summary output if an issue has no assignee.

Bump version to `0.2.2`.